f32 -> f64 global types

+++ src/global_types.rs
@@ -1,3 +1,3 @@
-pub type Pressure = f32;
-pub type Depth = f32;
-pub type Minutes = f32;
+pub type Pressure = f64;
+pub type Depth = f64;
+pub type Minutes = f64;
